insights: Refactor part 1 (insights-voyager)
These changes remove all the code no longer needed to run insights with
its own dedicated backend. Switching the backend to a new GrantNav
powered API.

This change adds to and maintains the existing functionality

This also adds the ability for custom dashboards such as
'cost of living' trackers to be spun up.